adjust stubs
